test: move test flowfile builder
