+++ tests/utils/builder/parallel.go
@@ -4,10 +4,6 @@ import (
 	"github.com/flowexec/flow/types/executable"
 )
 
-const (
-	parallelBaseDesc = "Multiple executables can be run concurrently using a parallel executable."
-)
-
 func ParallelExecByRefConfig(opts ...Option) *executable.Executable {
 	name := "parallel-config"
 	e1 := SimpleExec(opts...)
@@ -16,9 +12,6 @@ func ParallelExecByRefConfig(opts ...Option) *executable.Executable {
 		Verb:       "start",
 		Name:       name,
 		Visibility: privateExecVisibility(),
-		Description: parallelBaseDesc +
-			"The `execs` field can be used to define the child executables with more options. " +
-			"This includes defining an executable inline, retries, arguments, and more.",
 		Parallel: &executable.ParallelExecutableType{
 			Execs: []executable.ParallelRefConfig{
 				{Ref: e1.Ref()},
@@ -45,8 +38,6 @@ func ParallelExecWithExit(opts ...Option) *executable.Executable {
 		Name:       name,
 		Aliases:    []string{"parallel-exit"},
 		Visibility: privateExecVisibility(),
-		Description: parallelBaseDesc +
-			"The `failFast` option can be set to `true` to stop the flow if a sub-executable fails.",
 		Parallel: &executable.ParallelExecutableType{
 			FailFast: &ff,
 			Execs:    executable.ParallelRefConfigList{{Ref: e1.Ref()}, {Ref: e2.Ref()}, {Ref: e3.Ref()}},
@@ -61,8 +52,6 @@ func ParallelExecWithExit(opts ...Option) *executable.Executable {
 
 func ParallelExecWithMaxThreads(opts ...Option) *executable.Executable {
 	e := ParallelExecByRefConfig(opts...)
-	e.Description = parallelBaseDesc +
-		"\n\nThe `maxThreads` option can be set to limit the number of concurrent executions."
 	e.Parallel.MaxThreads = 1
 	return e
 }
