+++ tests/utils/builder/request.go
@@ -4,21 +4,12 @@ import (
 	"github.com/flowexec/flow/types/executable"
 )
 
-const (
-	requestBaseDesc = "Request executables send HTTP requests with the specified request and response settings.\n"
-)
-
 func RequestExec(opts ...Option) *executable.Executable {
 	name := "request"
-	docstring := requestBaseDesc +
-		"The `url` field is required and must be a valid URL. " +
-		"The `method` field is optional and defaults to `GET`.\n" +
-		"The `headers` field is optional and can be used to set request headers."
 	e := &executable.Executable{
-		Verb:        "run",
-		Name:        name,
-		Visibility:  privateExecVisibility(),
-		Description: docstring,
+		Verb:       "run",
+		Name:       name,
+		Visibility: privateExecVisibility(),
 		Request: &executable.RequestExecutableType{
 			URL:    "https://httpbin.org/get",
 			Method: "GET",
@@ -37,13 +28,10 @@ func RequestExec(opts ...Option) *executable.Executable {
 
 func RequestExecWithBody(opts ...Option) *executable.Executable {
 	name := "request-with-body"
-	docstring := requestBaseDesc +
-		"The `body` field is optional and can be used to send a request body."
 	e := &executable.Executable{
-		Verb:        "run",
-		Name:        name,
-		Visibility:  privateExecVisibility(),
-		Description: docstring,
+		Verb:       "run",
+		Name:       name,
+		Visibility: privateExecVisibility(),
 		Request: &executable.RequestExecutableType{
 			URL:    "https://httpbin.org/post",
 			Method: "POST",
@@ -59,13 +47,10 @@ func RequestExecWithBody(opts ...Option) *executable.Executable {
 
 func RequestExecWithTransform(opts ...Option) *executable.Executable {
 	name := "request-with-transform"
-	docstring := requestBaseDesc +
-		"The `transformResponse` field is optional and can be used to transform the response using an Expr expression."
 	e := &executable.Executable{
-		Verb:        "run",
-		Name:        name,
-		Visibility:  privateExecVisibility(),
-		Description: docstring,
+		Verb:       "run",
+		Name:       name,
+		Visibility: privateExecVisibility(),
 		Request: &executable.RequestExecutableType{
 			URL:               "https://httpbin.org/get",
 			TransformResponse: "status",
@@ -81,13 +66,10 @@ func RequestExecWithTransform(opts ...Option) *executable.Executable {
 
 func RequestExecWithTimeout(opts ...Option) *executable.Executable {
 	name := "request-with-timeout"
-	docstring := requestBaseDesc +
-		"The `timeout` field is optional and can be used to set the request timeout."
 	e := &executable.Executable{
-		Verb:        "run",
-		Name:        name,
-		Visibility:  privateExecVisibility(),
-		Description: docstring,
+		Verb:       "run",
+		Name:       name,
+		Visibility: privateExecVisibility(),
 		Request: &executable.RequestExecutableType{
 			URL:     "https://httpbin.org/delay/3",
 			Timeout: 1,
@@ -102,14 +84,10 @@ func RequestExecWithTimeout(opts ...Option) *executable.Executable {
 
 func RequestExecWithValidatedStatus(opts ...Option) *executable.Executable {
 	name := "request-with-validated-status"
-	docstring := requestBaseDesc +
-		"The `validStatusCodes` field is optional and can be used to specify the valid status codes. " +
-		"If the response status code is not in the list, the executable will fail."
 	e := &executable.Executable{
-		Verb:        "run",
-		Name:        name,
-		Visibility:  privateExecVisibility(),
-		Description: docstring,
+		Verb:       "run",
+		Name:       name,
+		Visibility: privateExecVisibility(),
 		Request: &executable.RequestExecutableType{
 			URL:              "https://httpbin.org/status/400",
 			ValidStatusCodes: []int{200},
