+++ tests/utils/builder/serial.go
@@ -4,10 +4,6 @@ import (
 	"github.com/flowexec/flow/types/executable"
 )
 
-const (
-	serialBaseDesc = "Multiple executables can be run in sequence using a serial executable.\n"
-)
-
 func SerialExecByRefConfig(opts ...Option) *executable.Executable {
 	name := "serial-config"
 	e1 := SimpleExec(opts...)
@@ -16,9 +12,6 @@ func SerialExecByRefConfig(opts ...Option) *executable.Executable {
 		Verb:       "start",
 		Name:       name,
 		Visibility: privateExecVisibility(),
-		Description: serialBaseDesc +
-			"The `execs` field can be used to define the child executables with more options. " +
-			"This includes defining an executable inline, retries, arguments, and more.",
 		Serial: &executable.SerialExecutableType{
 			Execs: []executable.SerialRefConfig{
 				{Ref: e1.Ref()},
@@ -45,8 +38,6 @@ func SerialExecWithExit(opts ...Option) *executable.Executable {
 		Name:       name,
 		Aliases:    []string{"serial-exit"},
 		Visibility: privateExecVisibility(),
-		Description: serialBaseDesc +
-			"The `failFast` option can be set to `true` to stop the executable if a sub-executable fails.",
 		Serial: &executable.SerialExecutableType{
 			FailFast: &ff,
 			Execs:    []executable.SerialRefConfig{{Ref: e1.Ref()}, {Ref: e2.Ref()}, {Ref: e3.Ref()}},
